+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/UriUtils.java
@@ -88,16 +88,35 @@ public final class UriUtils {
         String scheme = uri.getScheme();
         String path = uri.getPath();
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N && path != null) {
-            String[] externals = new String[]{"/external", "/external_path"};
+            String[] externals = new String[]{"/external/", "/external_path/"};
+            File file = null;
             for (String external : externals) {
-                if (path.startsWith(external + "/")) {
-                    File file = new File(Environment.getExternalStorageDirectory().getAbsolutePath()
-                            + path.replace(external, ""));
+                if (path.startsWith(external)) {
+                    file = new File(Environment.getExternalStorageDirectory().getAbsolutePath()
+                            + path.replace(external, "/"));
                     if (file.exists()) {
                         Log.d("UriUtils", uri.toString() + " -> " + external);
                         return file;
                     }
                 }
+            }
+            file = null;
+            if (path.startsWith("/files_path/")) {
+                file = new File(Utils.getApp().getFilesDir().getAbsolutePath()
+                        + path.replace("/files_path/", "/"));
+            } else if (path.startsWith("/cache_path/")) {
+                file = new File(Utils.getApp().getCacheDir().getAbsolutePath()
+                        + path.replace("/cache_path/", "/"));
+            } else if (path.startsWith("/external_files_path/")) {
+                file = new File(Utils.getApp().getExternalFilesDir(null).getAbsolutePath()
+                        + path.replace("/external_files_path/", "/"));
+            } else if (path.startsWith("/external_cache_path/")) {
+                file = new File(Utils.getApp().getExternalCacheDir().getAbsolutePath()
+                        + path.replace("/external_cache_path/", "/"));
+            }
+            if (file != null && file.exists()) {
+                Log.d("UriUtils", uri.toString() + " -> " + path);
+                return file;
             }
         }
         if (ContentResolver.SCHEME_FILE.equals(scheme)) {
@@ -278,10 +297,24 @@ public final class UriUtils {
 
     private static File copyUri2Cache(Uri uri) {
         Log.d("UriUtils", "copyUri2Cache() called");
-        InputStream is = uri2InputStream(uri);
-        File file = new File(Utils.getApp().getCacheDir(), "" + System.currentTimeMillis());
-        UtilsBridge.writeFileFromIS(file.getAbsolutePath(), is);
-        return file;
+        InputStream is = null;
+        try {
+            is = Utils.getApp().getContentResolver().openInputStream(uri);
+            File file = new File(Utils.getApp().getCacheDir(), "" + System.currentTimeMillis());
+            UtilsBridge.writeFileFromIS(file.getAbsolutePath(), is);
+            return file;
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+            return null;
+        } finally {
+            if (is != null) {
+                try {
+                    is.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
     }
 
     /**
@@ -290,14 +323,14 @@ public final class UriUtils {
      * @param uri The uri.
      * @return the input stream
      */
-    public static InputStream uri2InputStream(Uri uri) {
-        StringBuilder stringBuilder = new StringBuilder();
+    public static byte[] uri2Bytes(Uri uri) {
         InputStream is = null;
         try {
             is = Utils.getApp().getContentResolver().openInputStream(uri);
-            return is;
+            return UtilsBridge.inputStream2Bytes(is);
         } catch (FileNotFoundException e) {
             e.printStackTrace();
+            Log.d("UriUtils", "uri to bytes failed.");
             return null;
         } finally {
             if (is != null) {