+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/UtilsBridge.java
@@ -16,6 +16,7 @@ import com.google.gson.Gson;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.InputStream;
 import java.io.Serializable;
@@ -217,6 +218,10 @@ class UtilsBridge {
         return ConvertUtils.inputStream2Bytes(is);
     }
 
+    static ByteArrayOutputStream input2OutputStream(final InputStream is) {
+        return ConvertUtils.input2OutputStream(is);
+    }
+
     static List<String> inputStream2Lines(final InputStream is, final String charsetName) {
         return ConvertUtils.inputStream2Lines(is, charsetName);
     }
@@ -292,6 +297,10 @@ class UtilsBridge {
 
     static long getFsAvailableSize(String path) {
         return FileUtils.getFsAvailableSize(path);
+    }
+
+    static void notifySystemToScan(File file) {
+        FileUtils.notifySystemToScan(file);
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -380,6 +389,10 @@ class UtilsBridge {
         return IntentUtils.getSendSmsIntent(phoneNumber, content);
     }
 
+    static Intent getLaunchAppDetailsSettingsIntent(final String pkgName, final boolean isNewTask) {
+        return IntentUtils.getLaunchAppDetailsSettingsIntent(pkgName, isNewTask);
+    }
+
 
     ///////////////////////////////////////////////////////////////////////////
     // JsonUtils
@@ -405,6 +418,10 @@ class UtilsBridge {
     ///////////////////////////////////////////////////////////////////////////
     // PermissionUtils
     ///////////////////////////////////////////////////////////////////////////
+    static boolean isGranted(final String... permissions) {
+        return PermissionUtils.isGranted(permissions);
+    }
+
     @RequiresApi(api = Build.VERSION_CODES.M)
     static boolean isGrantedDrawOverlays() {
         return PermissionUtils.isGrantedDrawOverlays();
@@ -540,5 +557,9 @@ class UtilsBridge {
     ///////////////////////////////////////////////////////////////////////////
     static Uri file2Uri(final File file) {
         return UriUtils.file2Uri(file);
+    }
+
+    static File uri2File(final Uri uri) {
+        return UriUtils.uri2File(uri);
     }
 }